move overlay wrappers + cleanup
