+++ packages/renderer/src/screens/Action/Overlays/SignLedgerOverlay.tsx
@@ -9,7 +9,7 @@ import { useLedgerFeedback } from '@ren/contexts/action';
 import { useOverlay } from '@polkadot-live/ui/contexts';
 import { ButtonPrimary, ButtonSecondary } from '@polkadot-live/ui/kits/buttons';
 import { FlexColumn, FlexRow } from '@polkadot-live/ui/styles';
-import { LedgerOverlayWrapper } from '../Wrappers';
+import { LedgerOverlayWrapper } from './Wrappers';
 import { PuffLoader } from 'react-spinners';
 import {
   faChevronRight,
